requirements.txt
vllm
pandas
nlpstats
seaborn
matplotlib
transformers
numpy
unbabel-comet
tqdm

